+++ bank-account-backend/src/main/java/com/sg/kata/service/BankAccountService.java
@@ -2,10 +2,11 @@ package com.sg.kata.service;
 
 import com.sg.kata.dto.TransactionDto;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 public interface BankAccountService {
-    TransactionDto deposit(int amount);
-    TransactionDto withdraw(int amount);
+    TransactionDto deposit(BigDecimal amount);
+    TransactionDto withdraw(BigDecimal amount);
     List<TransactionDto> getStatement();
 }
